+++ mnist_snn/parameters.h
@@ -8,8 +8,6 @@
 #include <stdio.h>
 #include <stdlib.h>
 
-#include "utils.h"
-
 typedef enum
 {
     False,
@@ -26,35 +24,39 @@ int t_back = -5;
 int epoch = 1;
 
 // Input parameters
-double p_rest = -70.0;
+float p_rest = -70.0;
 
 // Neuron parameters
-double p_inhibit = -100.0;
-double p_th = -55.0;
-double p_hyperpolarization = -90.0;
-double p_drop = 0.8;
-double p_th_drop = 0.4;
+float p_inhibit = -100.0;
+float p_th = -55.0;
+float p_hyperpolarization = -90.0;
+float p_drop = 0.8;
+float p_th_drop = 0.4;
 
 double w_min = 0.00001;
-double w_max = 1; // np.max(synapse_init)
+double w_max = 1.0; // np.max(synapse_init)
 
 // STDP parameters
 double stdp_offset = 0.0;
 double sigma = 0.01;
 double A_plus = 0.8;
+// double A_plus = 0.6;
 double A_minus = 0.8;
+// double A_minus = 0.3;
 double tau_plus = 5.0;
+// double tau_plus = 8.0;
 double tau_minus = 5.0;
+// double tau_minus = 3.0;
 double mu = 0.9;
 
 // Receptive field parameters
-double w[5][5] = {{-0.5, -0.125, 0.25, -0.125, -0.25},
+float w[5][5] = {{-0.5, -0.125, 0.25, -0.125, -0.25},
                  {-0.125, 0.25, 0.625, 0.25, -0.125},
                  {0.25, 0.625, 1.0, 0.625, 0.25},
                  {-0.125, 0.25, 0.625, 0.25, -0.125},
                  {-0.5, -0.125, 0.25, -0.125, -0.5}};
 
-double ran[5] = {-2.0, -1.0, 0.0, 1.0, 2.0};
+float ran[5] = {-2.0, -1.0, 0.0, 1.0, 2.0};
 int ox = 2;
 int oy = 2;
